feat: remove SDKv2 stub and serve framework provider directly
- delete SDKv2 provider stub (mailgun/provider.go) and muxing layer
  (internal/framework/mux.go); main.go now serves the
  terraform-plugin-framework provider directly over protocol v6
- drop direct terraform-plugin-sdk/v2 and terraform-plugin-mux deps
- migrate acceptance tests to terraform-plugin-testing
- add .env loading for acceptance tests via godotenv in TestMain
  and gitignore .env
- move TestIsNotFound into mailgun/errors_test.go
