+++ .gitignore
@@ -26,6 +26,9 @@ website/node_modules
 
 website/vendor
 
+# Local environment files (acceptance test credentials)
+.env
+
 # Test exclusions
 !command/test-fixtures/**/*.tfstate
 !command/test-fixtures/**/.terraform/
