+++ internal/framework/acctest_helpers_test.go
@@ -1,28 +1,58 @@
 package framework_test
 
 import (
-	"context"
 	"fmt"
 	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+	"github.com/joho/godotenv"
 	"github.com/mailgun/mailgun-go/v5"
 
 	"github.com/wgebis/terraform-provider-mailgun/internal/framework"
 	mailgunpkg "github.com/wgebis/terraform-provider-mailgun/mailgun"
 )
 
-// protoV6Providers returns the muxed provider factory used by acceptance
-// tests in this package.
+// TestMain loads a local .env file (if present) before running the test suite
+// so acceptance tests can pick up MAILGUN_API_KEY without exporting it in the
+// shell. Existing environment variables always take precedence.
+func TestMain(m *testing.M) {
+	loadDotEnv()
+	os.Exit(m.Run())
+}
+
+// loadDotEnv searches the current directory and its parents for a .env file
+// and loads it. Missing files are ignored; this is a test-only convenience.
+func loadDotEnv() {
+	dir, err := os.Getwd()
+	if err != nil {
+		return
+	}
+	for {
+		candidate := filepath.Join(dir, ".env")
+		if _, err := os.Stat(candidate); err == nil {
+			_ = godotenv.Load(candidate)
+			return
+		}
+		parent := filepath.Dir(dir)
+		if parent == dir {
+			return
+		}
+		dir = parent
+	}
+}
+
+// protoV6Providers returns the provider factory used by acceptance tests in
+// this package.
 func protoV6Providers() map[string]func() (tfprotov6.ProviderServer, error) {
 	return map[string]func() (tfprotov6.ProviderServer, error){
 		"mailgun": func() (tfprotov6.ProviderServer, error) {
-			return framework.MuxedProviderServer(context.Background())
+			return framework.NewProviderServer()
 		},
 	}
 }
