+++ mailgun/config.go
@@ -1,24 +1,14 @@
 package mailgun
 
 import (
-	"log"
 	"strings"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/mailgun/mailgun-go/v5"
 )
 
 // Config struct holds API key
 type Config struct {
 	APIKey string
-}
-
-// Client returns a new client for accessing mailgun.
-func (c *Config) Client() (*Config, diag.Diagnostics) {
-
-	log.Printf("[INFO] Mailgun Client configured ")
-
-	return c, nil
 }
 
 // GetClient returns a fresh Mailgun client for the given region. A new client
