--- mailgun/mailgun_helper.go
@@ -1,35 +1,0 @@
-package mailgun
-
-import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"hash/crc32"
-	"strings"
-)
-
-func setDefaultRegionForImport(d *schema.ResourceData) {
-	parts := strings.SplitN(d.Id(), ":", 2)
-
-	if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-		_ = d.Set("region", "us")
-	} else {
-		_ = d.Set("region", parts[0])
-		d.SetId(parts[1])
-	}
-}
-
-// stringHashcode hashes a string to a unique hashcode.
-//
-// crc32 returns an uint32, but for our use we need
-// and non-negative integer. Here we cast to an integer
-// and invert it if the result is negative.
-func stringHashcode(s string) int {
-	v := int(crc32.ChecksumIEEE([]byte(s)))
-	if v >= 0 {
-		return v
-	}
-	if -v >= 0 {
-		return -v
-	}
-	// v == MinInt
-	return 0
-}
